fix bug: expire overflow change default channel add sd permission android:versionName="1.0.2.11" android:versionCode="12"
git-svn-id: https://hapiboke.googlecode.com/svn/trunk@11 6d090491-eb2e-0410-b838-ad2ac0d8293e

+++ podcast/xapppodcast/src/info/xuluan/podcast/service/PodcastService.java
@@ -56,11 +56,11 @@ public class PodcastService extends Service {
 
 	public int pref_connection_sel = WIFI_CONNECT;
 
-	public int pref_update_wifi = 0;
-	public int pref_update_mobile = 0;
-	public int pref_item_expire = 0;
-	public int pref_download_file_expire = 0;
-	public int pref_played_file_expire = 0;
+	public long pref_update_wifi = 0;
+	public long pref_update_mobile = 0;
+	public long pref_item_expire = 0;
+	public long pref_download_file_expire = 0;
+	public long pref_played_file_expire = 0;
 
 	private DownloadItemListener mDownloadListener = null;
